+++ src/components/auth/apple/AppleLogin.style.ts
@@ -1,0 +1,14 @@
+import {StyleSheet} from 'react-native';
+
+const appleLoginStyle = StyleSheet.create({
+  button: {
+    backgroundColor: 'black',
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingVertical: 12,
+    flexDirection: 'row',
+    borderRadius: 8,
+  },
+});
+
+export default appleLoginStyle;
